+++ R/dcem_star_train.R
@@ -45,6 +45,9 @@
 #'
 #'         \item (4) Priors: \strong{sample_out$prior}
 #'         A vector of priors.
+#'
+#'         \item (5) Membership: \strong{sample_out$membership}
+#'         A vector of cluster membership for data.
 #'         }
 #'
 #' @usage
@@ -81,6 +84,7 @@
 #' sample_mv_out$sigma
 #' sample_mv_out$prior
 #' sample_mv_out$prob
+#' print(sample_mv_out$membership)
 #'
 #' @author Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic
 #'
@@ -103,7 +107,6 @@ dcem_star_train <-
 
     if (missing(num_clusters)) {
       num_clusters = 2
-
       print("Using default value for number of clusters = 2.")
     }
     else{
@@ -119,7 +122,6 @@ dcem_star_train <-
       print("Using the improved Kmeans++ initialisation scheme.")
     }
 
-
     data <- apply(data, 2, as.numeric)
     data[is.na(data)] <- NULL
 
@@ -128,7 +130,7 @@ dcem_star_train <-
     num_data <- nrow(test_data)
     valid_columns <- ncol(test_data)
 
-    em_data_out = list()
+    emstar_out = list()
 
     if (valid_columns >= 2) {
 
@@ -145,7 +147,7 @@ dcem_star_train <-
       }
       sigma = sigma_mv(num_clusters, valid_columns)
       priors = get_priors(num_clusters)
-      em_data_out = dcem_star_cluster_mv(
+      emstar_out = dcem_star_cluster_mv(
         test_data,
         meu,
         sigma,
@@ -164,7 +166,7 @@ dcem_star_train <-
       }
       sigma = sigma_uv(test_data, num_clusters)
       priors = get_priors(num_clusters)
-      em_data_out = dcem_star_cluster_uv(
+      emstar_out = dcem_star_cluster_uv(
         test_data,
         meu,
         sigma,
@@ -174,5 +176,5 @@ dcem_star_train <-
         iteration_count)
     }
 
-    return(em_data_out)
+    return(emstar_out)
   }