+++ R/dcem_test.R
@@ -48,6 +48,9 @@ source("./R/dcem_train.R")
 #'
 #'         \item (4) Priors: \strong{sample_out$prior}
 #'         A vector of prior.
+#'
+#'         \item (5) Membership: \strong{membership}: A vector of cluster
+#'         membership for data.
 #'         }
 #'
 #' @usage